+++ packages/opencensus-core/test/test-tag-map.ts
@@ -16,6 +16,7 @@
 
 import * as assert from 'assert';
 import {TagMap} from '../src/tags/tag-map';
+import {TagTtl} from '../src/tags/types';
 
 describe('TagMap()', () => {
   let tagMap: TagMap;
@@ -25,18 +26,55 @@ describe('TagMap()', () => {
   const value1 = {value: 'value1'};
   const value2 = {value: 'value2'};
   const invalidValue1 = {value: 'a'.repeat(256)};
+  const NO_PROPAGATION_MD = {tagTtl: TagTtl.NO_PROPAGATION};
+  const UNLIMITED_PROPAGATION_MD = {tagTtl: TagTtl.UNLIMITED_PROPAGATION};
+
+  const expectedTagValueWithMetadata1 = {
+    tagValue: {value: 'value1'},
+    tagMetadata: {tagTtl: -1}
+  };
+
+  const expectedTagValueWithMetadata2 = {
+    tagValue: {value: 'value2'},
+    tagMetadata: {tagTtl: -1}
+  };
 
   beforeEach(() => {
     tagMap = new TagMap();
   });
 
   describe('set()', () => {
-    it('should set tagkey and tagvalue', () => {
+    it('should set tagkey and tagvalue with default TagMetadata', () => {
       tagMap.set(key1, value1);
-      const tags = tagMap.tags;
+      const tags = tagMap.tagsWithMetadata;
       assert.equal(tags.size, 1);
-      assert.deepStrictEqual(tags.get(key1), value1);
+      assert.deepStrictEqual(tags.get(key1), expectedTagValueWithMetadata1);
     });
+
+    it('should set tagkey and tagvalue with NO_PROPAGATION TagTtl', () => {
+      const expectedTagValueWithMetadata = {
+        tagValue: {value: 'value1'},
+        tagMetadata: {tagTtl: 0}
+      };
+
+      tagMap.set(key1, value1, NO_PROPAGATION_MD);
+      const tags = tagMap.tagsWithMetadata;
+      assert.equal(tags.size, 1);
+      assert.deepStrictEqual(tags.get(key1), expectedTagValueWithMetadata);
+    });
+
+    it('should set tagkey and tagvalue with UNLIMITED_PROPAGATION TagTtl',
+       () => {
+         const expectedTagValueWithMetadata = {
+           tagValue: {value: 'value1'},
+           tagMetadata: {tagTtl: -1}
+         };
+
+         tagMap.set(key1, value1, UNLIMITED_PROPAGATION_MD);
+         const tags = tagMap.tagsWithMetadata;
+         assert.equal(tags.size, 1);
+         assert.deepStrictEqual(tags.get(key1), expectedTagValueWithMetadata);
+       });
 
     it('should throw an error when invalid tagKey', () => {
       assert.throws(() => {
@@ -52,34 +90,34 @@ describe('TagMap()', () => {
 
     it('should not set duplicate tagkey and tagvalue', () => {
       tagMap.set(key1, value1);
-      const tags = tagMap.tags;
+      const tags = tagMap.tagsWithMetadata;
       assert.equal(tags.size, 1);
-      assert.deepStrictEqual(tags.get(key1), value1);
+      assert.deepStrictEqual(tags.get(key1), expectedTagValueWithMetadata1);
       tagMap.set(key1, value1);
       assert.equal(tags.size, 1);
     });
 
     it('should update existing tagkey', () => {
       tagMap.set(key1, value1);
-      const tags = tagMap.tags;
+      const tags = tagMap.tagsWithMetadata;
       assert.equal(tags.size, 1);
-      assert.deepStrictEqual(tags.get(key1), value1);
+      assert.deepStrictEqual(tags.get(key1), expectedTagValueWithMetadata1);
       tagMap.set(key1, value2);
       assert.equal(tags.size, 1);
-      assert.deepStrictEqual(tags.get(key1), value2);
+      assert.deepStrictEqual(tags.get(key1), expectedTagValueWithMetadata2);
     });
   });
   describe('delete()', () => {
     it('should delete tagkey', () => {
       tagMap.set(key1, value1);
-      const tags = tagMap.tags;
+      const tags = tagMap.tagsWithMetadata;
       assert.equal(tags.size, 1);
       tagMap.delete(key1);
       assert.equal(tags.size, 0);
     });
     it('should delete missing tagkey1', () => {
       tagMap.set(key1, value1);
-      const tags = tagMap.tags;
+      const tags = tagMap.tagsWithMetadata;
       assert.equal(tags.size, 1);
       tagMap.delete(key2);
       assert.equal(tags.size, 1);