+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - Add support for overriding sampling for a span.
 - Enforce `--strictNullChecks` and `--noUnusedLocals` Compiler Options on [opencensus-exporter-jaeger] packages.
 - Add support for recording Exemplars.
+- Add `TagMetadata` that defines the properties associated with a `Tag`.
 
 ## 0.0.9 - 2019-02-12
 - Add Metrics API.